テストが通るように設定とテストコードを追加
