+++ package.json
@@ -11,8 +11,10 @@
     "electron": "^18.3.7",
     "google-auth-library": "^7.14.0",
     "googleapis": "^95.0.0",
+    "jest": "^27.5.1",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "ts-jest": "^27.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.22.9",
@@ -58,13 +60,14 @@
     "yarn": "^1.22.11"
   },
   "scripts": {
+    "electron": "electron dist/main.js",
     "eslint": "eslint src/**/*.{js,jsx,ts,tsx}",
     "eslint:fix": "eslint src/**/*.{js,jsx,ts,tsx} --fix",
+    "test": "jest",
     "prettier": "prettier src/**/*.{js,jsx,ts,tsx} --check",
     "prettier:fix": "prettier src/**/*.{js,jsx,ts,tsx} --write",
     "webpack": "webpack --watch --progress",
     "webpack:server": "webpack server",
-    "webpack:build": "webpack --mode production",
-    "electron": "electron dist/main.js"
+    "webpack:build": "webpack --mode production"
   }
 }
